Test some more

+++ flopp-kernel/src/main/java/com/github/kjetilv/flopp/kernel/ByteSource.java
@@ -3,5 +3,5 @@ package com.github.kjetilv.flopp.kernel;
 @FunctionalInterface
 public interface ByteSource {
 
-    int fill(byte[] bytes, int offset, int length);
+    long fill(byte[] bytes, long offset, long length);
 }
